refactor: 애플리케이션 엔트리 포인트 변경
