+++ src/main/java/calculator/Application.java
@@ -1,11 +1,9 @@
 package calculator;
 
-import calculator.engine.ProcessController;
-import calculator.engine.RunningState;
+import calculator.engine.Controller;
 
 public class Application {
     public static void main(String[] args) {
-        ProcessController processController = new ProcessController();
-        processController.run(new RunningState(true));
+        new Controller().run();
     }
 }
