+++ src/main/java/calculator/Console.java
@@ -1,0 +1,4 @@
+package calculator;
+
+public class Console {
+}
