+++ src/main/java/calculator/engine/Controller.java
@@ -1,0 +1,9 @@
+package calculator.engine;
+
+public class Controller implements Runnable{
+
+    @Override
+    public void run() {
+
+    }
+}
