--- src/main/java/calculator/engine/ExpressionConverter.java
@@ -1,4 +1,0 @@
-package calculator.engine;
-
-public interface ExpressionConverter {
-}
