--- src/main/java/calculator/engine/PostfixConverter.java
@@ -1,5 +1,0 @@
-package calculator.engine;
-
-public class PostfixConverter implements ExpressionConverter{
-
-}
