+++ src/main/java/calculator/engine/converter/Converter.java
@@ -1,0 +1,4 @@
+package calculator.engine.converter;
+
+public interface Converter {
+}
