+++ src/main/java/calculator/engine/io/Characters.java
@@ -1,4 +1,4 @@
-package calculator.io;
+package calculator.engine.io;
 
 public enum Characters {
     BLANK(" "),
