+++ src/main/java/calculator/engine/io/Input.java
@@ -1,13 +1,13 @@
-package calculator.io;
+package calculator.engine.io;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class Reader {
+public class Input {
     private BufferedReader bufferedReader;
 
-    public Reader() {
+    public Input() {
         bufferedReader = new BufferedReader(new InputStreamReader(System.in));
     }
