+++ src/main/java/calculator/engine/io/SelectOption.java
@@ -1,4 +1,4 @@
-package calculator.io;
+package calculator.engine.io;
 
 import java.util.List;
