+++ src/main/java/calculator/engine/model/Expression.java
@@ -1,0 +1,4 @@
+package calculator.engine.model;
+
+public class Expression {
+}
