+++ src/main/java/calculator/engine/model/Operator.java
@@ -1,0 +1,5 @@
+package calculator.engine.model;
+
+public enum Operator {
+
+}
