+++ src/main/java/calculator/engine/repository/CalculationHistory.java
@@ -1,4 +1,4 @@
-package calculator.repository;
+package calculator.engine.repository;
 
 public class CalculationHistory implements History{
 }
