+++ src/main/java/calculator/engine/repository/History.java
@@ -1,0 +1,4 @@
+package calculator.engine.repository;
+
+public interface History {
+}
