--- src/main/java/calculator/model/Expression.java
@@ -1,4 +1,0 @@
-package calculator.model;
-
-public class Expression {
-}
