--- src/main/java/calculator/model/Operator.java
@@ -1,5 +1,0 @@
-package calculator.model;
-
-public enum Operator {
-
-}
