--- src/main/java/calculator/repository/History.java
@@ -1,4 +1,0 @@
-package calculator.repository;
-
-public interface History {
-}
